Width of mcm_scale<>() arg and return type depend on template args.
The input and return type of the mcm_scale<>() function are now tuned
according the to the number of bits specfied for the source and
destination. This avoids the caller having to perform error-prone
casts.

+++ include/midi2/ump_to_midi2.hpp
@@ -174,7 +174,7 @@ private:
         out0.bank = c.pn_msb;
         out0.index = c.pn_lsb;
         get<1>(out.w) =
-            mcm_scale<14, 32>((static_cast<std::uint32_t>(c.value_msb) << 7) | static_cast<std::uint32_t>(value));
+            mcm_scale<14, 32>(static_cast<std::uint16_t>((static_cast<std::uint16_t>(c.value_msb) << 7) | value));
         ctxt->push(out.w);
       }
     };
